frontend/requirements.txt
# Streamlit Frontend Requirements
# Compatible with Python 3.12

# Core Framework
streamlit==1.50.0

# HTTP Client
requests==2.32.5

# UI Components
plotly==6.3.1

# Environment
python-dotenv==1.2.1
